ray class, blended white and blue image

+++ raytracer/src/image/color.rs
@@ -1,3 +1,5 @@
 use crate::nalgebra::Vector3;
 
-pub type Color = Vector3<i64>;
+// TODO: should I just wrap the vector class to hide scale between [0,1] or [0,255]?
+/// Color as a unit vector
+pub type Color = Vector3<f64>;
